consolidate the factories
