+++ univocity-trader-core/pom.xml
@@ -19,7 +19,6 @@
 		<commons-codec.version>1.13</commons-codec.version>
 		<spring.version>5.2.2.RELEASE</spring.version>
 		<mysql.version>5.1.45</mysql.version>
-		<guice.version>4.2.2</guice.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -76,12 +75,7 @@
 			<artifactId>mail</artifactId>
 			<version>1.4.7</version>
 		</dependency>
-		
-		<dependency>
-			<groupId>com.google.inject</groupId>
-			<artifactId>guice</artifactId>
-			<version>${guice.version}</version>
-		</dependency>
+
 		<dependency>
 			<groupId>com.opencsv</groupId>
 			<artifactId>opencsv</artifactId>
