+++ univocity-trader-livetrader/src/main/java/com/univocity/trader/livetrader/LiveTraderRunner.java
@@ -12,32 +12,29 @@ import com.univocity.trader.account.OrderRequest;
 import com.univocity.trader.candles.Candle;
 import com.univocity.trader.config.UnivocityConfiguration;
 import com.univocity.trader.currency.Currencies;
-import com.univocity.trader.email.MailUtil;
 import com.univocity.trader.exchange.Exchange;
-import com.univocity.trader.exchange.ExchangeFactory;
 import com.univocity.trader.exchange.binance.api.client.domain.market.Candlestick;
-import com.univocity.trader.guice.UnivocityFactory;
+import com.univocity.trader.factory.UnivocityFactory;
 import com.univocity.trader.indicators.base.TimeInterval;
 import com.univocity.trader.notification.OrderExecutionToLog;
-import com.univocity.trader.strategy.StrategyFactory;
 
 public class LiveTraderRunner {
    public void run(String currenciesList) {
       final String[] currencies = Currencies.getInstance().fromList(currenciesList);
       final UnivocityConfiguration univocityConfiguration = UnivocityFactory.getInstance().getUnivocityConfiguration();
       System.out.println("Strategy: " + univocityConfiguration.getStrategyClass().getName());
-      final Exchange<Candlestick> exchange = ExchangeFactory.getInstance().getExchange(univocityConfiguration.getExchangeClass());
+      final Exchange<Candlestick> exchange = UnivocityFactory.getInstance().getExchange(univocityConfiguration.getExchangeClass());
       LiveTrader<Candlestick> liveTrader = null;
       try {
-         liveTrader = new LiveTrader<Candlestick>(exchange, TimeInterval.minutes(1), MailUtil.getEmailConfig());
+         liveTrader = new LiveTrader<Candlestick>(exchange, TimeInterval.minutes(1), UnivocityFactory.getInstance().getEmailConfig());
          final String apiKey = univocityConfiguration.getExchangeAPIKey();
          final String secret = univocityConfiguration.getExchangeAPISecret();
          final Client<?> client = liveTrader.addClient(univocityConfiguration.getExchangeClientId(), ZoneId.systemDefault(), "USDT", apiKey, secret);
          client.tradeWith(currencies);
-         client.strategies().add(StrategyFactory.getInstance().getStrategySupplier(univocityConfiguration.getStrategyClass()));
+         client.strategies().add(UnivocityFactory.getInstance().getStrategySupplier(univocityConfiguration.getStrategyClass()));
          for (final Class<?> clazz : univocityConfiguration.getStrategyMonitorClasses()) {
             System.out.println("Monitor: " + clazz.getName());
-            client.monitors().add(StrategyFactory.getInstance().getStrategyMonitorSupplier(clazz));
+            client.monitors().add(UnivocityFactory.getInstance().getStrategyMonitorSupplier(clazz));
          }
          client.account().maximumInvestmentAmountPerAsset(20);
          client.account().setOrderManager(new DefaultOrderManager() {
